refactor: Only modify page title on view action
(#18)

* Only modify page title on view action

Makes it so when you're viewing page history, editing, etc. it uses the default page title rather than the new one. This is something that was requested by my users that I hacked in.

* Update includes/WikiSEO.php

nice!

Co-authored-by: H. C. Kruse <6594492+octfx@users.noreply.github.com>

* Don't change the page title on diffs either

Actually hold on a second, isArticle() is true on diffs and that is one of the important cases where I wanted the default page title back. Is there a cleaner way of doing this?

* Update WikiSEO.php

* Update WikiSEO.php

* Update WikiSEO.php

does getRequest() need to be checked for null (?)

* Update includes/WikiSEO.php

works for me!

Co-authored-by: H. C. Kruse <6594492+octfx@users.noreply.github.com>

Co-authored-by: H. C. Kruse <6594492+octfx@users.noreply.github.com>
Change-Id: Ie1c9e9228f658e616d65b8619c13a82090361a1c