fix: Don't Save empty descriptions (#21)
* Don't save empty descriptions

Prevent completely empty descriptions from being saved to the database.

Before this it would generate an empty description page prop and show a blank description on page info (common for File pages or if the page has no intro). A blank meta description isn't even added to the HTML, so it has no use.

* Update DeferredDescriptionUpdate.php

So I assumed the delete had to be run in order to clear out any pre-existing blank ones in page props, but it wasn't needed, the blank ones cleared without it. Then I tried running without the delete at all and it seems to have no impact whatsoever on functionality, I don't think that is needed

Change-Id: I8a6e7de7fe83a79caf0c6486f6ca970f9ae0cbfc